# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Pet Management API
  description: Pet Management API with health logs, prescriptions, item master, and JWT authentication.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: '2'
servers:
  - url: http://localhost:9966/petmanagement/api/v2
...
security:
  - bearerAuth: []
paths:
...
  /visits:
    get:
      tags: [visit]
      operationId: listVisits
      summary: List visits with pagination (can filter by petId)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: petId
          in: query
          description: Filter visits by Pet ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitPage'
    post:
      tags: [visit]
      operationId: addVisit
      summary: Record a new visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
  /visits/{visitId}:
    get:
      tags: [visit]
      operationId: getVisit
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags: [visit]
      operationId: updateVisit
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitFields'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
    delete:
      tags: [visit]
      operationId: deleteVisit
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
  /visits/{visitId}/prescriptions:
    get:
      tags: [visit-prescription]
      operationId: listVisitPrescriptions
      summary: List prescriptions for a specific visit
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisitPrescription'
    post:
      tags: [visit-prescription]
      operationId: addVisitPrescription
      summary: Add a prescription to a visit
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitPrescriptionFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitPrescription'
  /visits/{visitId}/prescriptions/{visitPrescriptionId}:
    delete:
      tags: [visit-prescription]
      operationId: deleteVisitPrescription
      summary: Remove a prescription from a visit
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: visitPrescriptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
...
  /user-items:
    get:
      tags: [user-item]
      operationId: listUserItems
      summary: List user's items with pagination
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItemPage'
    post:
      tags: [user-item]
      operationId: addUserItem
      summary: Add a new user item record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserItemFields'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItem'
  /user-items/{userItemId}:
    get:
      tags: [user-item]
      operationId: getUserItem
      parameters:
        - name: userItemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItem'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    put:
      tags: [user-item]
      operationId: updateUserItem
      parameters:
        - name: userItemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserItemFields'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserItem'
    delete:
      tags: [user-item]
      operationId: deleteUserItem
      parameters:
        - name: userItemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: No Content
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Zero-based page index (0..N)
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: The size of the page to be returned
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
  schemas:
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          description: The size of the page
        totalElements:
          type: integer
          description: Total number of elements across all pages
        totalPages:
          type: integer
          description: Total number of pages
        number:
          type: integer
          description: Current page number (0-based)
      required:
        - size
        - totalElements
        - totalPages
        - number
...
    VisitPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Visit'
          required:
            - content
...
    UserItemPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/UserItem'
          required:
            - content
    ProblemDetail:
      title: Problem Detail
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        timestamp:
          type: string
          format: date-time
...
    VisitType:
      type: string
      enum: [vaccine, heartworm, flea_tick, checkup, general]
    Currency:
      type: string
      enum: [JPY, USD]
...
    VisitFields:
      title: Visit Fields
      type: object
      properties:
        petId:
          type: string
          format: uuid
        clinicId:
          type: string
          format: uuid
        visitedOn:
          type: string
          format: date
          example: '2025-05-20'
        weight:
          type: number
          format: float
          description: Weight in kg
          example: 5.50
        visitType:
          $ref: '#/components/schemas/VisitType'
        reason:
          type: string
          example: 'Annual vaccination'
        diagnosis:
          type: string
        treatment:
          type: string
        nextDueOn:
          type: string
          format: date
          example: '2026-05-20'
        totalFee:
          type: integer
          example: 5000
        currency:
          $ref: '#/components/schemas/Currency'
        note:
          type: string
      required:
        - petId
        - clinicId
        - visitedOn
    Visit:
      allOf:
        - $ref: '#/components/schemas/VisitFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - createdAt
            - updatedAt
    VisitPrescriptionFields:
      title: Visit Prescription Fields
      type: object
      properties:
        visitId:
          type: string
          format: uuid
        prescriptionId:
          type: string
          format: uuid
        quantity:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
          description: Quantity (e.g., 1.5)
          example: 1.0
        unit:
          type: string
          example: 'Tablet'
        days:
          type: integer
          minimum: 1
          example: 7
        dosageInstructions:
          type: string
          example: 'Once a day after meals'
        purpose:
          type: string
      required:
        - visitId
        - prescriptionId
        - quantity
        - unit
    VisitPrescription:
      allOf:
        - $ref: '#/components/schemas/VisitPrescriptionFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            visitId:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - visitId
            - createdAt
            - updatedAt
...
    UserItemFields:
      title: User Item Fields
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: The user ID
        itemId:
          type: string
          format: uuid
          description: The item ID
        note:
          type: string
          description: User's note for the item
        recordedAt:
          type: string
          format: date-time
          description: When the item was recorded by the user
      required:
        - userId
        - itemId
    UserItem:
      allOf:
        - $ref: '#/components/schemas/UserItemFields'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            createdAt:
              type: string
              format: date-time
              readOnly: true
            updatedAt:
              type: string
              format: date-time
              readOnly: true
          required:
            - id
            - createdAt
            - updatedAt